refactor(build): extract escHtml to shared/htmlEscape (DRY)
